\name{AMUN-package}
\alias{AMUN-package}
\alias{AMUN}
\docType{package}
\title{
\packageTitle{AMUN}
}
\description{
\packageDescription{AMUN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AMUN}
\packageIndices{AMUN}

}
\author{
\packageAuthor{AMUN}

Maintainer: \packageMaintainer{AMUN}
}
\references{
Morrison, D. F. (2005), Multivariate statistical methods, Series in Probability and Statistics, 4 edn, McGraw-Hill, New York.

Hardle, W. & Simar, L. (2007), Applied Multivariate Statistical Analysis,
Springer, New York.
}
\keyword{ package }
\seealso{

}
\examples{

}
