\name{AMUN-package}
\alias{AMUN-package}
\docType{package}
\title{
Analisis Multivariado
}
\description{
The packages presets useful tools for the Multivariate Analysis course at Universidad Nacional de Colombia, but it can be use by anybody who wants to computes some the examples of Morrison (2005) in chapters two, four and five 
}
\details{
This package performs some of the basics multivariate statistical analysis; they are programmed functions that are able to Test on means using the T square of hotelling, simultaneous intervals, the analysis of repeated measurements and groups of repeated measurements, profiles analysis, linear discriminate analysis and hypothesis over the variances and covariance matrices. Most of the functions uses the matrix of variance and covariance and not he data matrix since the aim of doing this package was originally to compute the class examples of the guide book Morrison (2005) and in some cases Simar (2007), where most of the time that matrix is shown.  Even though some function use the data matrix since it did not make sense no using other information.
\packageDESCRIPTION{AMUN}
\packageIndices{AMUN}

}
\author{
\packageAuthor{AMUN}

Maintainer: \packageMaintainer{AMUN}
}
\references{

}

\keyword{ package }
\seealso{
}
\examples{
}
