 \name{infarto}
 \alias{infarto}
 \title{Infarto}
 \description{Los datos corresponden a 99 personas v´ıctimas de infarto del miocardio (49 fallecieron (categoría 2) y 50 sobrevivieron (categoría 1)). Para cada una se midieron a la entrada del hospital 7 variables continuas.
  }
 \usage{data(infarto)}
 \format{Original data was read by using
 read.csv("infarto.csv", header = TRUE, sep = ",", row.names = 1)
}
 \source{}
 \references{}
  \keyword{datasets}